Refactor the shapes (naming, some internals) to be more self-contained instead of being linked to the button styles
