+++ Sources/SnappTheming/Builders/Foundation/EitherInsettableShape.swift
@@ -1,25 +1,11 @@
 //
-//  EitherShape.swift
+//  EitherInsettableShape.swift
 //  SnappTheming
 //
-//  Created by Ilian Konchev on 14.01.25.
+//  Created by Ilian Konchev on 15.01.25.
 //
 
 import SwiftUI
-
-public enum EitherShape<First: Shape, Second: Shape>: Shape {
-    case first(First)
-    case second(Second)
-    
-    public func path(in rect: CGRect) -> Path {
-        switch self {
-        case let .first(first):
-            return first.path(in: rect)
-        case let .second(second):
-            return second.path(in: rect)
-        }
-    }
-}
 
 public enum EitherInsettableShape<First: InsettableShape, Second: InsettableShape>: InsettableShape {
     case first(First)
