+++ Sources/SnappTheming/Builders/Foundation/EmptyShape.swift
