+++ Sources/SnappTheming/Theme/Shape/SnappThemingShapeDeclarations.swift
@@ -1,5 +1,5 @@
 //
-//  SnappThemingButtonStyleShapeDeclarations.swift
+//  SnappThemingShapeDeclarations.swift
 //  SnappTheming
 //
 //  Created by Oleksii Kolomiiets on 02.12.2024.
@@ -8,10 +8,10 @@
 import Foundation
 
 /// Manages button shape tokens, defining the appearance of button outlines, such as circular, rounded rectangle, or custom shapes.
-public typealias SnappThemingButtonStyleShapeDeclarations = SnappThemingDeclarations<SnappThemingButtonStyleShapeRepresentation, SnappThemingButtonStyleShapeConfiguration>
+public typealias SnappThemingShapeDeclarations = SnappThemingDeclarations<SnappThemingShapeRepresentation, SnappThemingShapeConfiguration>
 
-extension SnappThemingDeclarations where DeclaredValue == SnappThemingButtonStyleShapeRepresentation,
-                             Configuration == SnappThemingButtonStyleShapeConfiguration
+extension SnappThemingDeclarations where DeclaredValue == SnappThemingShapeRepresentation,
+                             Configuration == SnappThemingShapeConfiguration
 {
     public init(cache: [String: SnappThemingToken<DeclaredValue>]?, configuration: SnappThemingParserConfiguration = .default) {
         self.init(
